feat(fix-falcon-client-tests): mocks are ignored, but actually it is a fake, so renamed
